requirements.txt
anyio==3.6.1
bcrypt==4.0.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.85.0
greenlet==1.1.3
h11==0.14.0
idna==3.4
Naked==0.1.31
passlib==1.7.4
pycparser==2.21
pydantic==1.10.2
PyJWT==2.5.0
PyMySQL==1.0.2
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
shellescape==3.8.1
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
